ISSSUE-320 🚚 Move: SectionTitle 폴더 구조 변경
